+++ miden-tx/src/verifier/mod.rs
@@ -41,6 +41,7 @@ impl TransactionVerifier {
         let stack_outputs = TransactionKernel::build_output_stack(
             transaction.account_update().final_state_hash(),
             transaction.output_notes().commitment(),
+            transaction.expiration_block_num(),
         );
 
         // verify transaction proof