+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.6.0 (TBD)
 
+- Implemented kernel procedure to set transaction expiration block delta (#897).
+- Created a proving service that receives `TransactionWitness` and returns the proof using gRPC (#881).
 - Made note scripts public (#880).
 - Implemented serialization for `TransactionWitness`, `ChainMmr`, `TransactionInputs` and `TransactionArgs` (#888).
 - [BREAKING] Renamed the `TransactionProver` struct to `LocalTransactionProver` and added the `TransactionProver` trait (#865).
@@ -17,6 +19,8 @@
 - Implemented generation of transaction kernel procedure hashes in build.rs (#887).
 - [BREAKING] `send_asset` procedure was removed from the basic wallet (#829).
 - Implemented ability to invoke procedures against the foreign account (#882, #890, #896).
+- [BREAKING] Updated limits, introduced additional limits (#889).
+- Introduced `AccountDelta` maximum size limit of 32 KiB (#889).
 
 ## 0.5.1 (2024-08-28) - `miden-objects` crate only